+++ common_msgs/msg/TimePose.msg
@@ -1,3 +1,4 @@
 time timestamp
 geometry_msgs/Pose2D pose
 std_msgs/Int32 info
+std_msgs/Bool yorn
