+++ common_msgs/srv/hw3.srv
@@ -1,0 +1,4 @@
+int32 a
+int32 b
+---
+int32 sum
